requirements/dev.txt
appdirs==1.4.0
argon2-cffi==16.3.0
cffi==1.9.1
coverage==4.3.4
Django==1.10.5
django-crispy-forms==1.6.1
factory-boy==2.8.1
Faker==0.7.7
mock==2.0.0
olefile==0.44
packaging==16.8
pbr==2.0.0
Pillow==4.0.0
psycopg2==2.6.2
py==1.4.32
pycparser==2.17
pyparsing==2.1.10
pytest==3.0.6
pytest-django==3.1.2
pytest-sugar==0.8.0
python-dateutil==2.6.0
pytz==2016.10
selenium==3.0.2
six==1.10.0
termcolor==1.1.0
